Feat/Fix: Userlikes + friends endpoint works

+++ backend/Models/Friends.cs
@@ -2,10 +2,8 @@ namespace backend.Models
 {
     public class Friends
     {
-        public long UserId { get; set; }
-        public long UserFriendId { get; set; }
+        public int UserId { get; set; } 
+        public int UserFriendId { get; set; } 
         public DateTime Timestamp { get; set; }
     }
 }
-
-
